--- ui/src/KaptureApp/components/EditorSidebar/ComponentList/index.js
@@ -1,1 +1,0 @@
-export { ComponentList } from './connect';
