+++ ui/src/KaptureApp/components/EditorSidebar/Components/Components.module.css
@@ -1,4 +1,4 @@
-.componentList {
+.components {
   height: 100%;
   display: flex;
   flex-direction: column;
