+++ ui/src/KaptureApp/components/EditorSidebar/Components/ImportComponentsModal/ImportTextArea.js
