+++ ui/src/KaptureApp/components/EditorSidebar/Components/ImportComponentsModal/connect.js
