+++ ui/src/KaptureApp/components/EditorSidebar/Components/index.js
@@ -1,0 +1,1 @@
+export { Components } from './connect';
